+++ tests/test_clustering.py
@@ -33,7 +33,7 @@ import matplotlib.pyplot as plt
 from matplotlib.testing.decorators import image_comparison
 
 from sklearn.tree import DecisionTreeClassifier
-from sklearn.cluster import KMeans, MiniBatchKMeans
+from sklearn.cluster import KMeans, MiniBatchKMeans, SpectralClustering, BisectingKMeans
 from sklearn.datasets import load_iris as load_data
 from sklearn_evaluation import plot
 import sklearn_evaluation.plot.clustering as cl
@@ -62,6 +62,17 @@ def test_n_clusters_in_clf():
     clf = DummyClusterer()
     with pytest.raises(TypeError):
         plot.elbow_curve(X, clf)
+
+
+def test_score_in_clf_error():
+    clf = SpectralClustering()
+    with pytest.raises(AttributeError):
+        plot.elbow_curve(X, clf)
+
+
+@pytest.mark.parametrize("clf", [KMeans(), MiniBatchKMeans(), BisectingKMeans()])
+def test_score_methods_in_clf(clf):
+    plot.elbow_curve(X, clf, range_n_clusters=range(1, 10))
 
 
 def test_plot_elbow_curve_bad_input_value_error(ploomber_value_error_message):