+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 ## 0.10.0dev
-- [API Change] Fixed overlapping plot issue ([#211](https://github.com/ploomber/sklearn-evaluation/pull/211))
+- [API Change] A new figure and axes is created (via `plt.subplots()`) when calling a plotting method with `ax=None`. Previously, the current axes was used (via `plt.gca()`) ([#211](https://github.com/ploomber/sklearn-evaluation/pull/211))
 
 ## 0.9.0 (2023-01-13)