+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.10.0dev
 - [API Change] A new figure and axes is created (via `plt.subplots()`) when calling a plotting method with `ax=None`. Previously, the current axes was used (via `plt.gca()`) ([#211](https://github.com/ploomber/sklearn-evaluation/pull/211))
 - [Fix] Validating input elbow curve model has "score" method [#146] 
+- [Fix] Adds class labels for multi class roc plot (#209)
 
 ## 0.9.0 (2023-01-13)